+++ example/policy/package.install_always/xproto/xmake.lua
@@ -1,0 +1,6 @@
+add_rules("mode.debug", "mode.release")
+
+target("xproto", function()
+    set_kind("static")
+    add_files("xproto.cc")
+end)