+++ xmake/core/cache/master_misccache.lua
@@ -1,0 +1,4 @@
+-- "misc" is short for "miscellaneous". It is used to group items that don't belong to a specific category.
+-- For example, a list of miscellaneous tasks might include both administrative tasks and creative tasks,
+-- things that don't fall into one clear category.
+return require("cache/mastercache").cache("misc")