+++ xmake/core/sandbox/modules/import/core/cache/master_misccache.lua
@@ -1,0 +1,2 @@
+-- return module
+return require("cache/master_misccache")